+++ statistics/handle/handle.go
@@ -18,8 +18,8 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"github.com/pingcap/tidb/statistics/handle/extstats"
 	"math"
-	"slices"
 	"strconv"
 	"strings"
 	"sync"
@@ -47,7 +47,6 @@ import (
 	"github.com/pingcap/tidb/types"
 	"github.com/pingcap/tidb/util/chunk"
 	"github.com/pingcap/tidb/util/logutil"
-	"github.com/pingcap/tidb/util/mathutil"
 	"github.com/pingcap/tidb/util/sqlexec"
 	"github.com/tiancaiamao/gp"
 	"github.com/tikv/client-go/v2/oracle"
@@ -1147,157 +1146,22 @@ func (*Handle) releaseGlobalStatsReader(reader *statistics.StatsReader) error {
 
 // InsertExtendedStats inserts a record into mysql.stats_extended and update version in mysql.stats_meta.
 func (h *Handle) InsertExtendedStats(statsName string, colIDs []int64, tp int, tableID int64, ifNotExists bool) (err error) {
-	statsVer := uint64(0)
-	defer func() {
-		if err == nil && statsVer != 0 {
-			h.recordHistoricalStatsMeta(tableID, statsVer, StatsMetaHistorySourceExtendedStats)
-		}
-	}()
-	slices.Sort(colIDs)
-	bytes, err := json.Marshal(colIDs)
-	if err != nil {
-		return errors.Trace(err)
-	}
-	strColIDs := string(bytes)
-
-	se, err := h.pool.Get()
-	if err != nil {
-		return err
-	}
-	defer h.pool.Put(se)
-	exec := se.(sqlexec.SQLExecutor)
-	ctx := kv.WithInternalSourceType(context.Background(), kv.InternalTxnStats)
-
-	_, err = exec.ExecuteInternal(ctx, "begin pessimistic")
-	if err != nil {
-		return errors.Trace(err)
-	}
-	defer func() {
-		err = finishTransaction(ctx, exec, err)
-	}()
-	// No need to use `exec.ExecuteInternal` since we have acquired the lock.
-	rows, _, err := h.execRestrictedSQL(ctx, "SELECT name, type, column_ids FROM mysql.stats_extended WHERE table_id = %? and status in (%?, %?)", tableID, statistics.ExtendedStatsInited, statistics.ExtendedStatsAnalyzed)
-	if err != nil {
-		return errors.Trace(err)
-	}
-	for _, row := range rows {
-		currStatsName := row.GetString(0)
-		currTp := row.GetInt64(1)
-		currStrColIDs := row.GetString(2)
-		if currStatsName == statsName {
-			if ifNotExists {
-				return nil
-			}
-			return errors.Errorf("extended statistics '%s' for the specified table already exists", statsName)
-		}
-		if tp == int(currTp) && currStrColIDs == strColIDs {
-			return errors.Errorf("extended statistics '%s' with same type on same columns already exists", statsName)
-		}
-	}
-	version, err := getSessionTxnStartTS(se)
-	if err != nil {
-		return errors.Trace(err)
-	}
-	// Bump version in `mysql.stats_meta` to trigger stats cache refresh.
-	if _, err = exec.ExecuteInternal(ctx, "UPDATE mysql.stats_meta SET version = %? WHERE table_id = %?", version, tableID); err != nil {
-		return err
-	}
-	statsVer = version
-	// Remove the existing 'deleted' records.
-	if _, err = exec.ExecuteInternal(ctx, "DELETE FROM mysql.stats_extended WHERE name = %? and table_id = %?", statsName, tableID); err != nil {
-		return err
-	}
-	// Remove the cache item, which is necessary for cases like a cluster with 3 tidb instances, e.g, a, b and c.
-	// If tidb-a executes `alter table drop stats_extended` to mark the record as 'deleted', and before this operation
-	// is synchronized to other tidb instances, tidb-b executes `alter table add stats_extended`, which would delete
-	// the record from the table, tidb-b should delete the cached item synchronously. While for tidb-c, it has to wait for
-	// next `Update()` to remove the cached item then.
-	h.removeExtendedStatsItem(tableID, statsName)
-	const sql = "INSERT INTO mysql.stats_extended(name, type, table_id, column_ids, version, status) VALUES (%?, %?, %?, %?, %?, %?)"
-	if _, err = exec.ExecuteInternal(ctx, sql, statsName, tp, tableID, strColIDs, version, statistics.ExtendedStatsInited); err != nil {
-		return err
-	}
-	return
+	return h.callWithExec(func(sctx sessionctx.Context, exec sqlexec.RestrictedSQLExecutor) error {
+		return extstats.InsertExtendedStats(sctx, exec, h.recordHistoricalStatsMeta, h.updateStatsCache, h.statsCache.Load(), statsName, colIDs, tp, tableID, ifNotExists)
+	})
 }
 
 // MarkExtendedStatsDeleted update the status of mysql.stats_extended to be `deleted` and the version of mysql.stats_meta.
 func (h *Handle) MarkExtendedStatsDeleted(statsName string, tableID int64, ifExists bool) (err error) {
-	statsVer := uint64(0)
-	defer func() {
-		if err == nil && statsVer != 0 {
-			h.recordHistoricalStatsMeta(tableID, statsVer, StatsMetaHistorySourceExtendedStats)
-		}
-	}()
-	ctx := kv.WithInternalSourceType(context.Background(), kv.InternalTxnStats)
-	rows, _, err := h.execRestrictedSQL(ctx, "SELECT name FROM mysql.stats_extended WHERE name = %? and table_id = %? and status in (%?, %?)", statsName, tableID, statistics.ExtendedStatsInited, statistics.ExtendedStatsAnalyzed)
-	if err != nil {
-		return errors.Trace(err)
-	}
-	if len(rows) == 0 {
-		if ifExists {
-			return nil
-		}
-		return fmt.Errorf("extended statistics '%s' for the specified table does not exist", statsName)
-	}
-	if len(rows) > 1 {
-		logutil.BgLogger().Warn("unexpected duplicate extended stats records found", zap.String("name", statsName), zap.Int64("table_id", tableID))
-	}
-
-	se, err := h.pool.Get()
-	if err != nil {
-		return err
-	}
-	defer h.pool.Put(se)
-	exec := se.(sqlexec.SQLExecutor)
-
-	_, err = exec.ExecuteInternal(ctx, "begin pessimistic")
-	if err != nil {
-		return errors.Trace(err)
-	}
-	defer func() {
-		err1 := finishTransaction(ctx, exec, err)
-		if err == nil && err1 == nil {
-			h.removeExtendedStatsItem(tableID, statsName)
-		}
-		err = err1
-	}()
-	version, err := getSessionTxnStartTS(se)
-	if err != nil {
-		return errors.Trace(err)
-	}
-	if _, err = exec.ExecuteInternal(ctx, "UPDATE mysql.stats_meta SET version = %? WHERE table_id = %?", version, tableID); err != nil {
-		return err
-	}
-	statsVer = version
-	if _, err = exec.ExecuteInternal(ctx, "UPDATE mysql.stats_extended SET version = %?, status = %? WHERE name = %? and table_id = %?", version, statistics.ExtendedStatsDeleted, statsName, tableID); err != nil {
-		return err
-	}
-	return nil
+	return h.callWithExec(func(sctx sessionctx.Context, exec sqlexec.RestrictedSQLExecutor) error {
+		return extstats.MarkExtendedStatsDeleted(sctx, exec, h.recordHistoricalStatsMeta, h.updateStatsCache, h.statsCache.Load(), statsName, tableID, ifExists)
+	})
 }
 
 const updateStatsCacheRetryCnt = 5
 
-func (h *Handle) removeExtendedStatsItem(tableID int64, statsName string) {
-	for retry := updateStatsCacheRetryCnt; retry > 0; retry-- {
-		oldCache := h.statsCache.Load()
-		tbl, ok := oldCache.Get(tableID)
-		if !ok || tbl.ExtendedStats == nil || len(tbl.ExtendedStats.Stats) == 0 {
-			return
-		}
-		newTbl := tbl.Copy()
-		delete(newTbl.ExtendedStats.Stats, statsName)
-		if h.updateStatsCache(oldCache, []*statistics.Table{newTbl}, nil) {
-			return
-		}
-		if retry == 1 {
-			logutil.BgLogger().Info("remove extended stats cache failed", zap.String("stats_name", statsName), zap.Int64("table_id", tableID))
-		} else {
-			logutil.BgLogger().Info("remove extended stats cache failed, retrying", zap.String("stats_name", statsName), zap.Int64("table_id", tableID))
-		}
-	}
-}
-
 // ReloadExtendedStatistics drops the cache for extended statistics and reload data from mysql.stats_extended.
+// TODO: move this method to the `extstats` package.
 func (h *Handle) ReloadExtendedStatistics() error {
 	reader, err := h.getGlobalStatsReader(0)
 	if err != nil {
@@ -1328,177 +1192,20 @@ func (h *Handle) ReloadExtendedStatistics() error {
 
 // BuildExtendedStats build extended stats for column groups if needed based on the column samples.
 func (h *Handle) BuildExtendedStats(tableID int64, cols []*model.ColumnInfo, collectors []*statistics.SampleCollector) (*statistics.ExtendedStatsColl, error) {
-	ctx := kv.WithInternalSourceType(context.Background(), kv.InternalTxnStats)
-	const sql = "SELECT name, type, column_ids FROM mysql.stats_extended WHERE table_id = %? and status in (%?, %?)"
-	rows, _, err := h.execRestrictedSQL(ctx, sql, tableID, statistics.ExtendedStatsAnalyzed, statistics.ExtendedStatsInited)
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	if len(rows) == 0 {
-		return nil, nil
-	}
-	statsColl := statistics.NewExtendedStatsColl()
-	for _, row := range rows {
-		name := row.GetString(0)
-		item := &statistics.ExtendedStatsItem{Tp: uint8(row.GetInt64(1))}
-		colIDs := row.GetString(2)
-		err := json.Unmarshal([]byte(colIDs), &item.ColIDs)
-		if err != nil {
-			logutil.BgLogger().Error("invalid column_ids in mysql.stats_extended, skip collecting extended stats for this row", zap.String("column_ids", colIDs), zap.Error(err))
-			continue
-		}
-		item = h.fillExtendedStatsItemVals(item, cols, collectors)
-		if item != nil {
-			statsColl.Stats[name] = item
-		}
-	}
-	if len(statsColl.Stats) == 0 {
-		return nil, nil
-	}
-	return statsColl, nil
-}
-
-func (h *Handle) fillExtendedStatsItemVals(item *statistics.ExtendedStatsItem, cols []*model.ColumnInfo, collectors []*statistics.SampleCollector) *statistics.ExtendedStatsItem {
-	switch item.Tp {
-	case ast.StatsTypeCardinality, ast.StatsTypeDependency:
-		return nil
-	case ast.StatsTypeCorrelation:
-		return h.fillExtStatsCorrVals(item, cols, collectors)
-	}
-	return nil
-}
-
-func (h *Handle) fillExtStatsCorrVals(item *statistics.ExtendedStatsItem, cols []*model.ColumnInfo, collectors []*statistics.SampleCollector) *statistics.ExtendedStatsItem {
-	colOffsets := make([]int, 0, 2)
-	for _, id := range item.ColIDs {
-		for i, col := range cols {
-			if col.ID == id {
-				colOffsets = append(colOffsets, i)
-				break
-			}
-		}
-	}
-	if len(colOffsets) != 2 {
-		return nil
-	}
-	// samplesX and samplesY are in order of handle, i.e, their SampleItem.Ordinals are in order.
-	samplesX := collectors[colOffsets[0]].Samples
-	// We would modify Ordinal of samplesY, so we make a deep copy.
-	samplesY := statistics.CopySampleItems(collectors[colOffsets[1]].Samples)
-	sampleNum := mathutil.Min(len(samplesX), len(samplesY))
-	if sampleNum == 1 {
-		item.ScalarVals = 1
-		return item
-	}
-	if sampleNum <= 0 {
-		item.ScalarVals = 0
-		return item
-	}
-
-	se, seErr := h.pool.Get()
-	if seErr != nil {
-		logutil.BgLogger().Error("fail to get session", zap.String("category", "stats"), zap.Error(seErr))
-		return nil
-	}
-	defer h.pool.Put(se)
-	sctx := se.(sessionctx.Context)
-	sc := sctx.GetSessionVars().StmtCtx
-
+	var es *statistics.ExtendedStatsColl
 	var err error
-	samplesX, err = statistics.SortSampleItems(sc, samplesX)
-	if err != nil {
-		return nil
-	}
-	samplesYInXOrder := make([]*statistics.SampleItem, 0, sampleNum)
-	for i, itemX := range samplesX {
-		if itemX.Ordinal >= len(samplesY) {
-			continue
-		}
-		itemY := samplesY[itemX.Ordinal]
-		itemY.Ordinal = i
-		samplesYInXOrder = append(samplesYInXOrder, itemY)
-	}
-	samplesYInYOrder, err := statistics.SortSampleItems(sc, samplesYInXOrder)
-	if err != nil {
-		return nil
-	}
-	var corrXYSum float64
-	for i := 1; i < len(samplesYInYOrder); i++ {
-		corrXYSum += float64(i) * float64(samplesYInYOrder[i].Ordinal)
-	}
-	// X means the ordinal of the item in original sequence, Y means the oridnal of the item in the
-	// sorted sequence, we know that X and Y value sets are both:
-	// 0, 1, ..., sampleNum-1
-	// we can simply compute sum(X) = sum(Y) =
-	//    (sampleNum-1)*sampleNum / 2
-	// and sum(X^2) = sum(Y^2) =
-	//    (sampleNum-1)*sampleNum*(2*sampleNum-1) / 6
-	// We use "Pearson correlation coefficient" to compute the order correlation of columns,
-	// the formula is based on https://en.wikipedia.org/wiki/Pearson_correlation_coefficient.
-	// Note that (itemsCount*corrX2Sum - corrXSum*corrXSum) would never be zero when sampleNum is larger than 1.
-	itemsCount := float64(sampleNum)
-	corrXSum := (itemsCount - 1) * itemsCount / 2.0
-	corrX2Sum := (itemsCount - 1) * itemsCount * (2*itemsCount - 1) / 6.0
-	item.ScalarVals = (itemsCount*corrXYSum - corrXSum*corrXSum) / (itemsCount*corrX2Sum - corrXSum*corrXSum)
-	return item
+	err = h.callWithExec(func(sctx sessionctx.Context, exec sqlexec.RestrictedSQLExecutor) error {
+		es, err = extstats.BuildExtendedStats(sctx, exec, tableID, cols, collectors)
+		return err
+	})
+	return es, err
 }
 
 // SaveExtendedStatsToStorage writes extended stats of a table into mysql.stats_extended.
 func (h *Handle) SaveExtendedStatsToStorage(tableID int64, extStats *statistics.ExtendedStatsColl, isLoad bool) (err error) {
-	statsVer := uint64(0)
-	defer func() {
-		if err == nil && statsVer != 0 {
-			h.recordHistoricalStatsMeta(tableID, statsVer, StatsMetaHistorySourceExtendedStats)
-		}
-	}()
-	if extStats == nil || len(extStats.Stats) == 0 {
-		return nil
-	}
-
-	se, err := h.pool.Get()
-	if err != nil {
-		return err
-	}
-	defer h.pool.Put(se)
-	exec := se.(sqlexec.SQLExecutor)
-	ctx := kv.WithInternalSourceType(context.Background(), kv.InternalTxnStats)
-
-	_, err = exec.ExecuteInternal(ctx, "begin pessimistic")
-	if err != nil {
-		return errors.Trace(err)
-	}
-	defer func() {
-		err = finishTransaction(ctx, exec, err)
-	}()
-	version, err := getSessionTxnStartTS(se)
-	if err != nil {
-		return errors.Trace(err)
-	}
-	for name, item := range extStats.Stats {
-		bytes, err := json.Marshal(item.ColIDs)
-		if err != nil {
-			return errors.Trace(err)
-		}
-		strColIDs := string(bytes)
-		var statsStr string
-		switch item.Tp {
-		case ast.StatsTypeCardinality, ast.StatsTypeCorrelation:
-			statsStr = fmt.Sprintf("%f", item.ScalarVals)
-		case ast.StatsTypeDependency:
-			statsStr = item.StringVals
-		}
-		// If isLoad is true, it's INSERT; otherwise, it's UPDATE.
-		if _, err := exec.ExecuteInternal(ctx, "replace into mysql.stats_extended values (%?, %?, %?, %?, %?, %?, %?)", name, item.Tp, tableID, strColIDs, statsStr, version, statistics.ExtendedStatsAnalyzed); err != nil {
-			return err
-		}
-	}
-	if !isLoad {
-		if _, err := exec.ExecuteInternal(ctx, "UPDATE mysql.stats_meta SET version = %? WHERE table_id = %?", version, tableID); err != nil {
-			return err
-		}
-		statsVer = version
-	}
-	return nil
+	return h.callWithExec(func(sctx sessionctx.Context, exec sqlexec.RestrictedSQLExecutor) error {
+		return extstats.SaveExtendedStatsToStorage(sctx, exec, h.recordHistoricalStatsMeta, tableID, extStats, isLoad)
+	})
 }
 
 // CheckAnalyzeVersion checks whether all the statistics versions of this table's columns and indexes are the same.
